+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project somewhat adheres to [Semantic Versioning](https://semver.org/sp
 ### Fixed
 - Fix crash on older versions of Qt when reopening a project with certain windows open.
 - Fix potential crash when painting and the cursor leaves the map area.
+- Fix potential crash when changing maps with the Tileset Editor open.
 - Fix rare crash while quitting Porymap.
 - Fix `Edit > Clear Map Entries` in the Region Map Editor not saving the applied changes.
 - Fix `Edit > Undo/Redo` appearing enabled even when they don't do anything.